+++ src/main/kotlin/org/ionproject/integration/domain/common/Term.kt
@@ -1,4 +1,4 @@
-package org.ionproject.integration.domain
+package org.ionproject.integration.domain.common
 
 enum class Term(val number: Int) {
     FALL(1),
