+++ src/main/kotlin/org/ionproject/integration/domain/exception/TimetableTeachersBuilderException.kt
@@ -1,3 +1,3 @@
-package org.ionproject.integration.domain.exceptions
+package org.ionproject.integration.domain.exception
 
 class TimetableTeachersBuilderException(message: String) : Exception(message)
