--- src/main/kotlin/org/ionproject/integration/domain/timetable/Instructor.kt
@@ -1,5 +1,0 @@
-package org.ionproject.integration.domain.timetable
-
-data class Instructor(
-    val name: String
-)
