--- src/main/kotlin/org/ionproject/integration/domain/timetable/Label.kt
@@ -1,5 +1,0 @@
-package org.ionproject.integration.domain.timetable
-
-data class Label(
-    val acr: String
-)
