+++ src/main/kotlin/org/ionproject/integration/domain/timetable/dto/RawTimetableData.kt
@@ -1,4 +1,4 @@
-package org.ionproject.integration.domain.timetable
+package org.ionproject.integration.domain.timetable.dto
 
 data class RawTimetableData(
     val scheduleData: String,
