+++ src/main/kotlin/org/ionproject/integration/domain/timetable/model/ClassDetail.kt
@@ -1,4 +1,4 @@
-package org.ionproject.integration.domain.timetable
+package org.ionproject.integration.domain.timetable.model
 
 import java.lang.Integer.max
