+++ src/main/kotlin/org/ionproject/integration/domain/timetable/model/Course.kt
@@ -1,4 +1,4 @@
-package org.ionproject.integration.domain.timetable
+package org.ionproject.integration.domain.timetable.model
 
 data class Course(
     val label: Label,
