+++ src/main/kotlin/org/ionproject/integration/domain/timetable/model/CourseTeacher.kt
@@ -1,4 +1,6 @@
-package org.ionproject.integration.domain.timetable
+package org.ionproject.integration.domain.timetable.model
+
+import org.ionproject.integration.domain.common.School
 
 data class CourseTeacher(
     var school: School = School(),
