+++ src/main/kotlin/org/ionproject/integration/domain/timetable/model/EventCategory.kt
@@ -1,4 +1,4 @@
-package org.ionproject.integration.domain.timetable
+package org.ionproject.integration.domain.timetable.model
 
 enum class EventCategory {
     LECTURE,
