+++ src/main/kotlin/org/ionproject/integration/domain/timetable/model/Instructor.kt
@@ -1,0 +1,5 @@
+package org.ionproject.integration.domain.timetable.model
+
+data class Instructor(
+    val name: String
+)
