+++ src/main/kotlin/org/ionproject/integration/domain/timetable/model/Label.kt
@@ -1,0 +1,5 @@
+package org.ionproject.integration.domain.timetable.model
+
+data class Label(
+    val acr: String
+)
