+++ src/main/kotlin/org/ionproject/integration/domain/timetable/model/Programme.kt
@@ -1,4 +1,4 @@
-package org.ionproject.integration.domain.timetable
+package org.ionproject.integration.domain.timetable.model
 
 data class Programme(
     var name: String = "",
