+++ src/main/kotlin/org/ionproject/integration/infrastructure/repository/IntegrationJobRepository.kt
@@ -2,7 +2,7 @@ package org.ionproject.integration.infrastructure.repository
 
 import org.ionproject.integration.application.JobEngine
 import org.ionproject.integration.infrastructure.OutputFormat
-import org.ionproject.integration.ui.dto.JobType
+import org.ionproject.integration.application.job.JobType
 import org.springframework.jdbc.core.JdbcTemplate
 import org.springframework.jdbc.core.ResultSetExtractor
 import org.springframework.stereotype.Repository
