+++ src/main/kotlin/org/ionproject/integration/ui/dto/InputProcessor.kt
@@ -1,12 +1,11 @@
 package org.ionproject.integration.ui.dto
 
 import org.ionproject.integration.application.JobEngine
-import org.ionproject.integration.infrastructure.OutputFormat
+import org.ionproject.integration.application.job.JobType
 import org.ionproject.integration.infrastructure.exceptions.ArgumentException
 import org.ionproject.integration.infrastructure.repository.IInstitutionRepository
 import org.ionproject.integration.infrastructure.repository.IProgrammeRepository
 import org.springframework.stereotype.Component
-import java.lang.UnsupportedOperationException
 
 @Component
 class InputProcessor(
@@ -24,81 +23,14 @@ class InputProcessor(
 
         return dto.toSafeDto(type)
     }
-}
 
-enum class JobType(val identifier: String) {
-    TIMETABLE("timetable"),
-    EXAM_SCHEDULE("evaluation"),
-    ACADEMIC_CALENDAR("calendar");
+    private fun validateDto(dto: CreateJobDto, jobType: JobType) {
+        runCatching {
+            requireNotNull(dto.format) { "format" }
+            requireNotNull(dto.institution) { "institution" }
 
-    companion object Factory {
-        fun of(name: String?): JobType? = values().find { it.identifier.equals(name?.trim(), ignoreCase = true) }
+            if (jobType == JobType.TIMETABLE || jobType == JobType.EXAM_SCHEDULE)
+                requireNotNull(dto.programme) { "programme" }
+        }.onFailure { throw ArgumentException("Parameter '${it.message}' missing") }
     }
-}
-
-sealed interface SafeJobDto {
-    val institution: String
-    val format: String
-
-    fun toJobRequest(
-        institutionRepo: IInstitutionRepository,
-        programmeRepo: IProgrammeRepository
-    ): JobEngine.AbstractJobRequest
-}
-
-data class CreateJobDto(
-    val institution: String? = null,
-    val programme: String? = null,
-    val format: String? = null,
-    val type: String? = null
-) {
-    fun toSafeDto(type: JobType): SafeJobDto {
-        return when (type) {
-            JobType.TIMETABLE -> SafeTimetableJobDto(institution!!, programme!!, format!!)
-            JobType.ACADEMIC_CALENDAR -> SafeCalendarJobDto(institution!!, format!!)
-            JobType.EXAM_SCHEDULE -> throw UnsupportedOperationException("Exam Schedule not yet supported")
-        }
-    }
-}
-
-data class SafeTimetableJobDto(
-    override val institution: String,
-    val programme: String,
-    override val format: String,
-) : SafeJobDto {
-    override fun toJobRequest(
-        institutionRepo: IInstitutionRepository,
-        programmeRepo: IProgrammeRepository
-    ): JobEngine.AbstractJobRequest {
-        val format = OutputFormat.of(format)
-        val institution = institutionRepo.getInstitutionByIdentifier(institution)
-        val programme = programmeRepo.getProgrammeByAcronymAndInstitution(programme, institution)
-
-        return JobEngine.TimetableJobRequest(format, institution, programme)
-    }
-}
-
-data class SafeCalendarJobDto(
-    override val institution: String,
-    override val format: String,
-) : SafeJobDto {
-    override fun toJobRequest(
-        institutionRepo: IInstitutionRepository,
-        programmeRepo: IProgrammeRepository
-    ): JobEngine.AbstractJobRequest {
-        val format = OutputFormat.of(format)
-        val institution = institutionRepo.getInstitutionByIdentifier(institution)
-
-        return JobEngine.CalendarJobRequest(format, institution)
-    }
-}
-
-private fun validateDto(dto: CreateJobDto, jobType: JobType) {
-    runCatching {
-        requireNotNull(dto.format) { "format" }
-        requireNotNull(dto.institution) { "institution" }
-
-        if (jobType == JobType.TIMETABLE || jobType == JobType.EXAM_SCHEDULE)
-            requireNotNull(dto.programme) { "programme" }
-    }.onFailure { throw ArgumentException("Parameter '${it.message}' missing") }
 }
