+++ src/main/kotlin/org/ionproject/integration/ui/dto/SafeCalendarJobDto.kt
@@ -1,0 +1,21 @@
+package org.ionproject.integration.ui.dto
+
+import org.ionproject.integration.application.JobEngine
+import org.ionproject.integration.infrastructure.OutputFormat
+import org.ionproject.integration.infrastructure.repository.IInstitutionRepository
+import org.ionproject.integration.infrastructure.repository.IProgrammeRepository
+
+data class SafeCalendarJobDto(
+    override val institution: String,
+    override val format: String,
+) : SafeJobDto {
+    override fun toJobRequest(
+        institutionRepo: IInstitutionRepository,
+        programmeRepo: IProgrammeRepository
+    ): JobEngine.AbstractJobRequest {
+        val format = OutputFormat.of(format)
+        val institution = institutionRepo.getInstitutionByIdentifier(institution)
+
+        return JobEngine.CalendarJobRequest(format, institution)
+    }
+}
