+++ src/main/kotlin/org/ionproject/integration/ui/dto/SafeJobDto.kt
@@ -1,0 +1,15 @@
+package org.ionproject.integration.ui.dto
+
+import org.ionproject.integration.application.JobEngine
+import org.ionproject.integration.infrastructure.repository.IInstitutionRepository
+import org.ionproject.integration.infrastructure.repository.IProgrammeRepository
+
+interface SafeJobDto {
+    val institution: String
+    val format: String
+
+    fun toJobRequest(
+        institutionRepo: IInstitutionRepository,
+        programmeRepo: IProgrammeRepository
+    ): JobEngine.AbstractJobRequest
+}
