+++ src/test/kotlin/org/ionproject/integration/step/tasklet/iseltimetable/implementations/MappingTaskletTest.kt
@@ -5,11 +5,11 @@ import org.ionproject.integration.IOnIntegrationApplication
 import org.ionproject.integration.application.job.tasklet.MappingTasklet
 import org.ionproject.integration.job.ISELTimetableJob
 import org.ionproject.integration.job.TIMETABLE_JOB_NAME
-import org.ionproject.integration.domain.timetable.EventCategory
-import org.ionproject.integration.domain.timetable.Instructor
+import org.ionproject.integration.domain.timetable.model.EventCategory
+import org.ionproject.integration.domain.timetable.model.Instructor
 import org.ionproject.integration.domain.timetable.TimetableTeachers
-import org.ionproject.integration.domain.timetable.Weekday
-import org.ionproject.integration.domain.timetable.RawTimetableData
+import org.ionproject.integration.domain.common.Weekday
+import org.ionproject.integration.domain.timetable.dto.RawTimetableData
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
